chore!: Remove open keyword from Noir (#4967)
Removes the unused `open` keyword from Noir, as requested
[here](https://github.com/AztecProtocol/aztec-packages/pull/4898#issuecomment-1978566866).

**Breaking change**: This PR changes the format of Noir contract build
artifacts, so a full rebuild of contracts is needed after this PR is
merged. No other action is needed by the user. If they fail to do it,
they'll see a message along the lines of _No custom attributes found for
contract function foo. Try rebuilding the contract with the latest nargo
version._ (cc @AztecProtocol/devrel @rahul-kothari)

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>