test: refactor public cross chain tests for speed (#6082)
Fixing #6055, time goes from 400 seconds -> 150 seconds.

Minor issues when dealing with the snapshots across instances as the
cross chain harness is taking the pxe and node as args for its
constructor so if going directly from snapshot could run into issue
where it is connected to a different instance or the like.

Also ran into a deserialization issue for the aztec address, unclear why
as it seems like it have been registered 🤷