fix: shared mutable private getter fixes (#6652)
When I was making some changes to the SharedMutablePrivateGetter, I ran
into some issues with it.

The first was that the context variable was not being correctly updated
/ propogated due to not passing it as a mutable reference. The second
was that the SharedMutablePrivateGetter was reading the delay of slot of
its own address, and not reading the address of the specific other
contract to read from. The value was reading correctly but the delay was
not. I added two tests, one as a basic one to test max_block_number
propogation, and the other to test changing delays in SharedMutable, and
seeing whether the SharedMutablePrivateGetter would accurately reflect
those changes.