feat: add flag to not mask address in siloing enc log (#6668)
In some cases, encrypted event logs are intended to be broadcast with
their contract address publicly known.

To hide the address, we by default mask it (`masked_contract_address =
pedersen(addr, randomness)`), then use it to silo encrypted event logs.
To allow an app to optionally reveal the address by siloing the log with
it unmasked, it can provide a `randomness` value of 0.

Added docs, a test in private tail, and modified the existing tests in
`e2e_block_building` to include this case.