chore: Ultra flavor cleanup (#7070)
Removes a ton of duplication from ultra and ultra recursive flavors.
This had the side effect of slightly changing the order of one of the
get_all methods which required updates to the combiner test suite. The
suite has been updated to include one test that depends on the
"hand-computable" python values as originally intended, and one that
checks the combiner optimization consistency without dependence on
python generated values.

---------

Co-authored-by: Rumata888 <isennovskiy@gmail.com>