refactor: proper portal setup for fees + test (#7944)
Fixes #7692.

Gets rid of assets directly in the rollup contract. The rollup contract
will instead make a call to the portal to get it to pay some funds to
the sequencer, currently named `distributeFees`.

Calling initialize on the fee juice portal require that it have been
funded to ensure that the deployment of the l2 fee assets does not lead
to an undercollateralized case.