feat: automate verify_honk_proof input generation (#8092)
This script replaces the gen_inner_proof_inputs_ultra_honk.sh with
regenerate_verify_honk_proof_inputs.sh, which does the resetting of
test_programs and input copying automatically so that people don't have
to know intricate details on what to do to regenerate the inputs for the
verify_honk_proof test program.

As part of this, I created a python script
barretenberg/acir_tests/update_verify_honk_proof_inputs.py, which does
the copying/surgery of the proofs. This python script shouldn't have to
be run by itself.

It also updates the rebuild.sh script to actually output which
test_programs failed. Before, it always outputted that the rebuild
succeeded... It also allows for users to specify which test_programs
that they want to rebuild.

I updated the reset_acir_tests.sh script to be more efficient as well.