fix: Do not reuse anvil admin key (#8304)
In #8277 we added a PUBLIC_API_KEY that was used for accessing anvil
with cheat codes disabled, and the API_KEY was left for "admin" access
with cheat codes allowed. However, the API_KEY is also needed for
connecting to the node, which is is needed for retrieving proofs, needed
for prover nodes to operate.

This PR changes how keys are handled. API_KEY is now mapped to the anvil
public access, so it is not disclosed when sharing access to the node
API. A new FORK_ADMIN_API_KEY is introduced for admin access, and only
used for bootstrapping operations and for the bootstrapping sequencer.