feat: new test programs for wasm benchmarking (#8389)
This PR creates new test programs: bench_2_to_17, which is just a
program constructed out of poseidon hashes with around 2^17 - epsilon
gates, fold_2_to_17, which is a program that has 2 2^17-eps sized
circuits that would get folded together, and single_verify_proof, which
is a plonk recursive verifier which now has 2^18+eps gates.

It also reworks some of the interfaces in main.ts. In particular, it
gets the foldAndVerifyProgram flow working with fold_basic (and
SMALL_TEST execution trace structure).

I used the fold_2_to_17 test program to benchmark memory usage of
ClientIVC in WASM, which came out to be 700MiB. Note that this was only
possible by turning off the structure, because it would fail otherwise
from too many poseidon gates.

Running ClientIVC on the fold_basic test program using
TraceStructure::SMALL_TEST (which ends up with dyadic size of 2^18)
gives 1166.56MiB in WASM. Note that the builder memory is not an actual
full 2^18 gate circuit because the circuits only had 22, 4539, 16432
gates, so it should actually be close to 1250MiB or so if we had full
2^18 gate circuits.