test: add sync suite (#8550)
Fixes #8536 by introducing an initial setup for testing synching.

Adds a new "spam" contract that uses a large amount of data, such that
we can see the differences between "noop" transactions and more complex
things.

Adds a new `l1StartTime` option to the setup such that we can do an
instant "warp" to make the execution of past stored blocks much simpler.

While the intention was to have a lot of blocks to see synching over a
longer period, building that with non-trivial transactions will take
until I retire.
![Aging Matt Damon
GIF](https://media4.giphy.com/media/GrUhLU9q3nyRG/giphy.gif)

The test is **NOT** set to run in CI explicitly as it is a "measure"
test more than anything else, and without there being a target it is a
bit weird to restrict it 🤷