feat: reduce max memory in translator by freeing accumulator and eccvm (#8253)
We free the accumulator as soon as the decider_prover is done, the
eccvm_builder as soon as the eccvm_prover is constructed, the
translator_builder as soon as the translator_prover is constructed, and
the eccvm_prover, as soon as we pass all the necessary information to
the translator_builder. The peak memory drops by 18MB, but what's
significant is that the translator is no longer the bottleneck of the
benchmark, mostly due to freeing the accumulator.

The peak memory during the eccvm/translator drops from 666MiB to 328MiB,
a drop of 338MiB, which is approximately the size of one instance and
the commitment key, which is as expected.
Before:
<img width="1180" alt="Screenshot 2024-09-20 at 1 55 13 PM"
src="https://github.com/user-attachments/assets/469fe1e0-01ac-4e4e-a794-0266d15b4a5f">

After:
<img width="1193" alt="Screenshot 2024-09-19 at 6 28 39 PM"
src="https://github.com/user-attachments/assets/2f5c8873-5aa4-4631-a6f1-9a01aeef633c">