chore: big synching case + stability (#9022)
- Moves `cheat_code` log from `info` to `verbose`
- Hardens fixture creation,
- Create fixture for 1000 blocks of 4 private transfers
- Some plumbing to make it work

Hardens the fixture creation by introducing a "timejump" before the
first L2 block will land. By making the timejump a fixed jump ahead of a
starting time, we can ensure that we have some fixed upper limit and as
long as we are performing the setup within it we should be good.
In this case, that period is 200 blocks.

Beware, that the `Rollup` contract MUST be deployed at the same time,
since the `GENESIS_TIME` is used for all future slot computations, e.g.,
messing with that is pain. So if possible, please only change add stuff
to the L1 deployments AFTER the rollup deployment.

Since the big fixture of 1000 blocks is big and our system is very slow,
it will be skipped unless the `RUN_THE_BIG_ONE` flag is set.

```
aztec:e2e_synching [INFO] Stats: blockCount: 1000, txCount: 4, complexity: PrivateTransfer: {"numberOfBlocks":1008,"syncTime":225.7148497581482} +4m
```