fix: remove reliance on invalid decompositions in selector calculation (#9337)
Decomposing `hash` into `SELECTOR_SIZE` bytes relies on the compiler not
checking the correctness of decompositions performed at compile time.
This means the current code does not work at runtime and we're also
looking to enforce the correctness of this decomposition at
compile-time.

I've replaced this decomposition with a truncation into 32 bits as this
seems equivalent.