feat(avm): more efficient low leaf search (#9870)
This swaps out the initial implementation of the tracking just the
`indexedTreeMin` of a leaf seen by an indexed tree in favour of a vector
of sorted keys that we binary search over. This has 2 benefits,
especially in the scenario where the `indexedTreeMin` was storing very
small key values.

1) logn search time for keys that we have already seen
2) better than linear searching of the merkleDB for keys we dont have,
since we dont have to start searching from min