chore(avm): more pilcom compat changes (#10569)
More changes that I need for compatibility between old and new witgen.

In particular, I'm using `bb::{{vm_name}}` namespaces versus
``{{name}}Composer/Verifier`` etc because it makes it easier to work
with both of them. Also I had to add a prefix to the macros because
namespaces don't work with them.