fix: Archiver does not jump the gun on epoch completed (#10801)
Archiver's isEpochComplete would return `true` when the first L1 block
of the last epoch L2 slot was mined, ignoring the fact that the L2 block
could be posted on a later L1 slot.

This fixes the check so it reports the epoch as completed only once the
last L1 block of the last L2 slot of the epoch is mined.

Fixes #10800