fix: Faster polling times for archiver and sequencer (#11262)
When timetable is enforced, the combined polling times of the archiver
and sequencer could add to just over 3s, causing the sequencer to miss
its slot, since we require that block building starts within the first
3s. See
[here](https://github.com/AztecProtocol/aztec-packages/actions/runs/12806915326/job/35707395434?pr=11256#step:3:956)
for an example run on CI where this happened.

This PR changes the default polling from 1s to 500ms for archiver and
sequencer, forces 200ms for the block-building e2e test that depends on
the timetable, and updates the p2p client so it properly gets its config
rather than extracting it from env directly.