fix: Idempotent deploy-l1-contracts with initial validators (#11284)
Setting initial validators was causing deploy-l1-contracts to fail when
called a second time over the same set of L1 contracts. This PR checks
which validators of the initial ones have been already registered and
skips them.

Also fixes the typings for the `l1Artifacts` so they are not generic
ABIs and we get type-checking over the contract methods while deploying.
This caused viem to start complaining about using an empty array to
signal no args in calls like `contract.write.method([])`, which needed a
few fixes all over the place.

Fixes #11283