feat: new transcript functionality to hash elements without including in proof (#12233)
Adds a new public add_to_hash_buffer function to the transcript that
adds elements to the current_round_data buffer without adding them to
the proof_data. This will enable us to hash things without explicitly
including it in the proof like vk data that needs to be Fiat-Shamired
but doesn't need to be sent in the proof.