feat: precomputed ClientIVC VKs (#12126)
After this PR, we no longer rely on a user provided vk when verifying
civc proofs. This was insecure.

`yarn-project/bb-prover` now has a `yarn generate` step that creates two
files in `yarn-project/bb-prover/artifacts`: `private-civc-vk` and
`public-civc-vk`. These correspond to clientivc stacks that end in the
private and public tail, respectively.

This is achieved by pinning historic CIVC inputs and using one public
and private example, respectively. If the number of public inputs in the
tail circuits, or the fundamental structure, change we will need to bump
this. This pinning will be obsoleted by
https://github.com/AztecProtocol/barretenberg/issues/1296.

The write_vk command **is to be considered undocumented**. It is subject
to change. Namely, a future simplification
(https://github.com/AztecProtocol/barretenberg/issues/1296) will make it
not take an ivc stack. Original comments by Cody below:
```
Add a write_vk command to generate a vk for generating ClientIVC proofs. This consists of: a vk for 'the' hiding circuit, a vk for the ECCVM and a vk for the Translator. The later two could and perhaps should go away in the future since they are actually just fixed constants known at C++ compile time. The former sounds like a constant, but in fact the key depends on the number of outputs of the final circuit in a stack to be verified. At the moment the two possibilities in our system are the private kernel tail and tail-to-public circuits, where the latter I'm told has very many PIs, enough that we should have a distinction. I believe this means having two Tubes, or making the Tube receive exeuction time input on which of the two keys it should use.

We remove the special handling of single circuits in ClientIVC. This was originally added so that there would be _some_ unit tests of the bb binary of ClientIVC, but the new tests in this PR will fill that role better by being more realistic.
I also shove some little API improvements requested by @saleel in here to make sure they don't get lost in the shuffle.
```

Closes https://github.com/AztecProtocol/barretenberg/issues/1245

---------

Co-authored-by: ludamad <domuradical@gmail.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>