chore: separated multi map and fixed indexeddb map (#12896)
This PR cleans up `kv-store` a bit, attempting to separate maps and
multimaps a little bit better. The old approach caused a very stupid but
difficult to track bug in the `IndexedDB` implementation that would
cause partial notes to never be discovered and simulations to crash.

Also removed the `WithSize` variants of maps that were not implemented
in most store types and had a different testing flow, since they appear
to be completely unused @Maddiaa0

@alexghr do you think we can get rid of the old lmdb implementations
soon?